[TASK] Split SysFileReferenceLocalizedParentExists
Check SysFileReferenceLocalizedParentExists is a bit
risky and we will work on this to communicate better
with another patch.

However, sys_file_reference records that are
soft-deleted, and sys_language_uid > 0 and have an
invalid l10n_parent can be removed more safely since
they are deleted=1 and thus don't impact BE / FE.

We'll split those off to an own check before
SysFileReferenceLocalizedParentExists.