[FEATURE] Update l10n_state if translated values do not match
In case TCA allowLanguageSynchronization=1 is set the value of
a translated record should be the same as the value in the
default language, unless l10n_state contains a configuration
for the field != "parent" (e.g. "custom").

In this case the value is inconsistent because "Use value from
default language" is displayed in the BE, but the value is
actually different.

For better consistency, we change the configuration in l10n_state
for these cases to "custom".

Resolves: #97