[TASK] Add `typo3/cms-install` as dev dependency
The `setup` command is part of `EXT:install` and
required to setup instances, for example for the
`cli` phpt based testsuite.

This system extension is kind of optional and no
longer installed by default in composer instances,
but literally needed when the setup command should
be used to setup instances. Add it as development
dependency !

```shell
composer req --dev --no-update \
  "typo3/cms-install":"^12.4 || ^13.4"
```