[BUGFIX] Handle records having l10n_parent = uid
TcaTablesTranslatedParentInvalidPointer stumbles when it encounters
a records having "transOrigPointerField" (l10n_parent) set to its
own uid. It also stumbles when an l10n_parent of a different row
points to such a record.

To fix this, two new checks are added: Scan for such records in
pages and set them deleted, then scan for such records in all other
tables and set them deleted.

Additionally, sanitize TcaTablesTranslatedParentInvalidPointer to
skip such records which may happen if the check has to run multiple
times in one run because it follows a chain.