[BUGFIX] Deal with different pid / l10n_parent in localizations
When a localization is first changed in workspaces and after
that the default language record is moved in workspaces,
l10n_parent of the workspace-moved localization record
overlay points to live. PagesTranslatedLanguageParentDifferentPid
and TcaTablesTranslatedLanguageParentDifferentPid needs to deal
with this odd core behavior to avoid deleting such records.

Related: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89803
Resolves: #152